Breaking up components into individual files
