+++ huxley/www/js/components/RegistrationView.js
@@ -9,32 +9,32 @@ var cx = require('classnames');
 var React = require('react');
 var ReactRouter = require('react-router');
 
-var Button = require('components/Button');
 var CommitteeStore = require('stores/CommitteeStore');
 var ContactTypes = require('constants/ContactTypes');
 var ConferenceContext = require('components/ConferenceContext');
 var CountrySelect = require('components/CountrySelect');
 var CountryStore = require('stores/CountryStore');
 var GenderConstants = require('constants/GenderConstants');
-var NavLink = require('components/NavLink');
-var NumberInput = require('components/NumberInput');
 var OuterView = require('components/OuterView');
-var PhoneInput = require('components/PhoneInput');
 var ProgramTypes = require('constants/ProgramTypes');
+var RegistrationAccountInformation = require('components/RegistrationAccountInformation');
+var RegistrationComments = require('components/RegistrationComments');
+var RegistrationCountryPreferences = require('components/RegistrationCountryPreferences');
+var RegistrationFooter = require('components/RegistrationFooter');
+var RegistrationHeader = require('components/RegistrationHeader');
+var RegistrationPhoneInput = require('components/RegistrationPhoneInput');
+var RegistrationPrimaryContact = require('components/RegistrationPrimaryContact');
+var RegistrationProgramInformation = require('components/RegistrationProgramInformation');
+var RegistrationSchoolInformation = require('components/RegistrationSchoolInformation');
+var RegistrationSecondaryContact = require('components/RegistrationSecondaryContact');
+var RegistrationSpecialCommitteePreferences = require('components/RegistrationSpecialCommitteePreferences');
 var ServerAPI = require('lib/ServerAPI');
 var StatusLabel = require('components/StatusLabel');
-var TextInput = require('components/TextInput');
-var TextTemplate = require('components/TextTemplate');
 var _handleChange = require('utils/_handleChange');
 
 require('css/RegistrationView.less');
-var RegistrationViewText = require('text/RegistrationViewText.md');
 
 var USA = 'United States of America';
-
-var _accessSafe = function(obj, key) {
-  return obj && obj[key];
-};
 
 var RegistrationView = React.createClass({
   mixins: [
@@ -236,7 +236,7 @@ var RegistrationView = React.createClass({
         <form id="registration" onSubmit={this._handleSubmit}>
           <RegistrationHeader session={conference.session} />
           <hr />
-          <AccountInformation 
+          <RegistrationAccountInformation 
             handlers={accountInformationHandlers} 
             errors={accountInformationErrors} 
             accountInformation={accountInformationValues} 
@@ -244,7 +244,7 @@ var RegistrationView = React.createClass({
             focus={this._handlePasswordFocus}
           />
           <hr />
-          <SchoolInformation  
+          <RegistrationSchoolInformation  
             handlers={schoolInformationHandlers} 
             errors={schoolInformationErrors} 
             schoolInformation={schoolInformationValues} 
@@ -252,7 +252,7 @@ var RegistrationView = React.createClass({
             schoolInternational={this.state.school_international} 
           />
           <hr />
-          <ProgramInformation 
+          <RegistrationProgramInformation 
             handlers={programInformationHandlers} 
             errors={programInformationErrors} 
             programInformation={programInformationValues} 
@@ -260,7 +260,7 @@ var RegistrationView = React.createClass({
             programType={this.state.program_type}
           />
           <hr />
-          <PrimaryContact 
+          <RegistrationPrimaryContact 
             handlers={primaryContactHandlers} 
             primaryContactInformation={primaryContactValues} 
             errors={primaryContactErrors} 
@@ -269,7 +269,7 @@ var RegistrationView = React.createClass({
             isInternational={this.state.school_international} 
           />
           <hr />
-          <SecondaryContact 
+          <RegistrationSecondaryContact 
             handlers={secondaryContactHandlers} 
             secondaryContactInformation={secondaryContactValues} 
             errors={secondaryContactErrors} 
@@ -278,25 +278,25 @@ var RegistrationView = React.createClass({
             isInternational={this.state.school_international}
           />
           <hr />
-            <CountryPreferences
-              renderCountryDropdown={this.renderCountryDropdown}
-            />
+          <RegistrationCountryPreferences
+            renderCountryDropdown={this.renderCountryDropdown}
+          />
           <hr />
-            <SpecialCommitteePreferences 
-              handlers={specialCommitteeHandlers} 
-              errors={specialCommitteeErrors}
-              specialCommitteePrefValues={specialCommitteeValues} 
-              renderCommittees={this.renderCommittees}
-            />
+          <RegistrationSpecialCommitteePreferences 
+            handlers={specialCommitteeHandlers} 
+            errors={specialCommitteeErrors}
+            specialCommitteePrefValues={specialCommitteeValues} 
+            renderCommittees={this.renderCommittees}
+          />
           <hr />
-            <Comments
-              handler={_handleChange.bind(this, 'registration_comments')} 
-              value={this.state.registration_comments} 
-            />
+          <RegistrationComments
+            handler={_handleChange.bind(this, 'registration_comments')} 
+            value={this.state.registration_comments} 
+          />
           <hr />
-            <RegistrationFooter 
-              loading={this.state.loading} 
-            />
+          <RegistrationFooter 
+            loading={this.state.loading} 
+          />
         </form>
       </OuterView>
     );
@@ -569,528 +569,6 @@ var RegistrationView = React.createClass({
       this.context.shake && this.context.shake();
     });
   }
-});
-
-const RegistrationTextInput = React.createClass({
-  propTypes: {
-    errors: React.PropTypes.arrayOf(React.PropTypes.string),
-    onChange: React.PropTypes.func,
-    placeholder: React.PropTypes.string,
-    value: React.PropTypes.string,
-    type: React.PropTypes.oneOf(['text', 'password']),
-  },
-
-  render() {
-    const {errors, ...inputProps} = this.props;
-    return (
-      <div className="reg-field">
-        <TextInput {...inputProps} />
-        {errors && errors.map(error =>
-          <StatusLabel status="error">{error}</StatusLabel>
-        )}
-      </div>
-    );
-  },
-});
-
-const RegistrationPhoneInput = React.createClass({
-  propTypes: {
-    errors: React.PropTypes.arrayOf(React.PropTypes.string),
-    onChange: React.PropTypes.func,
-    placeholder: React.PropTypes.string,
-    value: React.PropTypes.string,
-  },
-
-  render() {
-    const {errors, ...inputProps} = this.props;
-    return (
-      <div className="reg-field">
-        <PhoneInput {...inputProps} />
-        {errors && errors.map(error =>
-          <StatusLabel status="error">{error}</StatusLabel>
-        )}
-      </div>
-    );
-  },
-});
-
-const RegistrationHeader = React.createClass({
-  propTypes: {
-    session: React.PropTypes.number,
-  },
-
-  render: function() {
-    return (
-      <div>
-        <TextTemplate conferenceSession={this.props.session}>
-          {RegistrationViewText}
-        </TextTemplate>
-        <NavLink direction="left" href="/login">
-          Back to Login
-        </NavLink>
-      </div>
-    );
-  },
-});
-
-const AccountInformation = React.createClass({
-  propTypes: {
-    handlers: React.PropTypes.object,
-    errors: React.PropTypes.object,
-    accountInformation: React.PropTypes.object,
-    blur: React.PropTypes.func,
-    focus: React.PropTypes.func,
-  },
-
-  render: function() {
-    var accessErrors = _accessSafe.bind(this, this.props.errors);
-    var accessHandlers = _accessSafe.bind(this, this.props.handlers);
-    var accessAccount = _accessSafe.bind(this, this.props.accountInformation);
-    return (
-      <div id="account_information">
-        <h3>Account Information</h3>
-        <RegistrationTextInput  
-          errors={accessErrors('first_name')} 
-          placeholder="First Name" 
-          onChange={accessHandlers('first_name')} 
-          value={accessAccount('first_name')} 
-        />
-        <RegistrationTextInput 
-          errors={accessErrors('last_name')} 
-          placeholder="Last Name" 
-          onChange={accessHandlers('last_name')} 
-          value={accessAccount('last_name')} 
-        />
-        <RegistrationTextInput
-          errors={accessErrors('username')} 
-          placeholder="Username" 
-          onChange={accessHandlers('username')} 
-          value={accessAccount('username')}
-        />
-        <RegistrationTextInput
-          errors={accessErrors('password')} 
-          type="password" 
-          placeholder="Password" 
-          value={accessAccount('password')} 
-          onChange={accessHandlers('password')}
-          onBlur={this.props.blur} 
-          onFocus={this.props.focus} 
-        />
-        <RegistrationTextInput
-          errors={accessErrors('password_confirm')} 
-          type="password" 
-          placeholder="Password (confirm)" 
-          value={accessAccount('password_confirm')} 
-          onChange={accessHandlers('password_confirm')} 
-          onBlur={this.props.blur}
-          onFocus={this.props.focus} 
-        />
-      </div>
-    );
-  },
-
-});
-
-const SchoolInformation = React.createClass({
-  propTypes: {
-    handlers: React.PropTypes.object,
-    errors: React.PropTypes.object,
-    schoolInformation: React.PropTypes.object,
-    handleInternationalChange: React.PropTypes.func,
-    schoolInternational: React.PropTypes.bool,
-  },
-
-  render: function() {
-    var accessErrors = _accessSafe.bind(this, this.props.errors);
-    var accessHandlers = _accessSafe.bind(this, this.props.handlers);
-    var accessSchool = _accessSafe.bind(this, this.props.schoolInformation);
-    return (
-      <div id="school_information">
-        <h3>School Information</h3>
-        <p className="instructions">Where is your school located?</p>
-        <ul>
-          <li>
-            <label>
-              <input 
-                type="radio" 
-                value=''
-                onChange={this.props.handleInternationalChange} 
-                checked={!this.props.schoolInternational}
-              /> United States of America
-            </label>
-          </li>
-          <li>
-            <label>
-              <input 
-                type="radio" 
-                value='international'
-                onChange={this.props.handleInternationalChange} 
-                checked={this.props.schoolInternational}
-              /> International
-            </label>
-          </li>
-        </ul>
-        <RegistrationTextInput 
-          errors={accessErrors('school_name')} 
-          placeholder="Official School Name" 
-          onChange={accessHandlers('school_name')} 
-          value={accessSchool('school_name')} 
-        />
-        <RegistrationTextInput  
-          errors={accessErrors('school_address')} 
-          placeholder="Street Address" 
-          onChange={accessHandlers('school_address')} 
-          value={accessSchool('school_address')} 
-        />
-        <RegistrationTextInput  
-          errors={accessErrors('school_city')} 
-          placeholder="City" 
-          onChange={accessHandlers('school_city')} 
-          value={accessSchool('school_city')} 
-        />
-        <RegistrationTextInput  
-          errors={accessErrors('school_state')} 
-          placeholder="State" 
-          onChange={accessHandlers('school_state')} 
-          value={accessSchool('school_state')} 
-        />
-        <RegistrationTextInput  
-          errors={accessErrors('school_zip')} 
-          placeholder="Zip" 
-          onChange={accessHandlers('school_zip')} 
-          value={accessSchool('school_zip')} 
-        />
-        <RegistrationTextInput  
-          errors={accessErrors('school_country')} 
-          placeholder="Country" 
-          onChange={accessHandlers('school_country')} 
-          value={accessSchool('school_country')} 
-          disabled={!this.props.schoolInternational} 
-          isControlled={true} 
-        />
-      </div>
-    );
-  },
-
-});
-
-const ProgramInformation = React.createClass({
-  propTypes: {
-    handlers: React.PropTypes.object,
-    errors: React.PropTypes.object,
-    programInformation: React.PropTypes.object,
-    handleProgramTypeChange: React.PropTypes.func,
-    programType: React.PropTypes.oneOf([ProgramTypes.CLUB,ProgramTypes.CLASS]),
-  },
-
-  render: function() {
-    var accessHandlers = _accessSafe.bind(this, this.props.handlers);
-    var accessErrors = _accessSafe.bind(this, this.props.errors);
-    var accessProgram = _accessSafe.bind(this, this.props.programInformation);
-    return (
-      <div id="program_information">
-        <h3>Program Information</h3>
-        <p className="instructions">What category best describes your program?</p>
-        <ul>
-          <li>
-            <label>
-              <input 
-                type="radio" 
-                checked={this.props.programType == ProgramTypes.CLUB} 
-                value={ProgramTypes.CLUB} 
-                onChange={this.props.handleProgramTypeChange}
-              /> Club
-            </label>
-          </li>
-          <li>
-            <label>
-              <input 
-                type="radio" 
-                value={ProgramTypes.CLASS} 
-                checked={this.props.propTypes == ProgramTypes.CLASS} 
-                onChange={this.props.handleProgramTypeChange} 
-              /> Class
-            </label>
-          </li>
-        </ul>
-        <p className="instructions">Please tell us a bit more about your delegation this
-           year. Provide us with the tentative number of beginner,
-           intermediate, and advanced delegates you intend to bring to BMUN.
-           Try to provide us with realistic estimates for your delegate
-           numbers in each category so we can provide your delegation with
-           the appropriate number and type of assignments.</p>
-        <NumberInput 
-          placeholder="Number of BMUN sessions attended" 
-          onChange={accessHandlers('times_attended')} 
-          value={accessProgram('times_attended')} 
-        />
-        {accessErrors('times_attended')}
-        <NumberInput 
-          placeholder="Tentative Number of Beginner Delegates" 
-          onChange={accessHandlers('beginner_delegates')} 
-          value={accessProgram('beginner_delegates')}
-        />
-        <label className="hint">
-          Beginner: Attended 0-3 conferences, not very familiar with Model
-          United Nations.
-        </label>
-        {accessErrors('beginner_delegates')}
-        <NumberInput 
-          placeholder="Tentative Number of Intermediate Delegates" 
-          onChange={accessHandlers('intermediate_delegates')} 
-          value={accessProgram('intermediate_delegates')} 
-        />
-        <label className="hint">
-          Intermediate: Attended 4-7 conferences, little to no practice in
-          advanced committees.
-        </label>
-        {accessErrors('intermediate_delegates')}
-        <NumberInput 
-          placeholder="Tentative Number of Advanced Delegates" 
-          onChange={accessHandlers('advanced_delegates')} 
-          value={accessProgram('advanced_delegates')} 
-        />
-        <label className="hint">
-          Advanced: Attended more than seven conferences, has participated
-          in many diverse committees.
-        </label>
-        {accessErrors('advanced_delegates')}
-        <p className="instructions">
-          Tentative Total Number of Delegates: {this._handleDelegateSum(
-            accessProgram('beginner_delegates'), accessProgram('intermediate_delegates'), 
-            accessProgram('advanced_delegates')
-          )}
-        </p>
-      </div>
-    );
-  },
-
-  _handleDelegateSum: function(beginner, intermediate, advanced) {
-    var sum = 0;
-    if (beginner) {
-      sum += parseInt(beginner, 10) || 0;
-    } if (intermediate) {
-      sum += parseInt(intermediate, 10) || 0; 
-    } if (advanced) {
-      sum += parseInt(advanced, 10) || 0;
-    }
-    return sum;
-  }
-
-});
-
-const PrimaryContact = React.createClass({
-  propTypes: {
-    handlers: React.PropTypes.object,
-    primaryContactInformation: React.PropTypes.object,
-    errors: React.PropTypes.object,
-    renderContactGenderField: React.PropTypes.func,
-    renderContactTypeField: React.PropTypes.func,
-    isInternational: React.PropTypes.bool,
-  },
-
-  render: function() {
-    var accessHandlers = _accessSafe.bind(this, this.props.handlers);
-    var accessPrimary = _accessSafe.bind(this, this.props.primaryContactInformation);
-    var accessErrors = _accessSafe.bind(this, this.props.errors);
-    return (
-      <div id='primary_contact'>
-        <h3>Primary Contact</h3>
-        {this.props.renderContactGenderField('primary_gender')}
-        <RegistrationTextInput
-          errors={accessErrors('primary_name')} 
-          placeholder="Name" 
-          onChange={accessHandlers('primary_name')} 
-          value={accessPrimary('primary_name')} 
-        />
-        <RegistrationTextInput 
-          errors={accessErrors('primary_email')} 
-          placeholder="Email" 
-          onChange={accessHandlers('primary_email')} 
-          value={accessPrimary('primary_email')} 
-        />
-        <RegistrationTextInput 
-          errors={accessErrors('primary_phone')} 
-          onChange={accessHandlers('primary_phone')} 
-          value={accessPrimary('primary_phone')} 
-          placeholder="Phone Number" 
-          isInternational={this.props.isInternational} 
-        />
-        {this.props.renderContactTypeField('primary_type')} 
-      </div>
-    );
-  },
-});
-
-const SecondaryContact = React.createClass({
-  propTypes: {
-    handlers: React.PropTypes.object,
-    secondaryContactInformation: React.PropTypes.object,
-    errors: React.PropTypes.object,
-    renderContactGenderField: React.PropTypes.func,
-    renderContactTypeField: React.PropTypes.func,
-    isInternational: React.PropTypes.bool,
-  },
-
-  render: function() {
-    var accessHandlers = _accessSafe.bind(this, this.props.handlers);
-    var accessSecondary = _accessSafe.bind(this, this.props.secondaryContactInformation);
-    var accessErrors = _accessSafe.bind(this, this.props.errors)
-    return (
-      <div id='secondary_contact'>
-        <h3>Secondary Contact</h3>
-        {this.props.renderContactGenderField('secondary_gender')}
-        <RegistrationTextInput
-          errors={accessErrors('secondary_name')}
-          placeholder="Name" 
-          onChange={accessHandlers('secondary_name')}
-          value={accessSecondary('secondary_name')}
-        />
-        <RegistrationTextInput 
-          errors={accessErrors('secondary_email')}
-          placeholder="Email" 
-          onChange={accessHandlers('secondary_email')}
-          value={accessSecondary('secondary_email')}
-        />
-        <RegistrationTextInput 
-          errors={accessErrors('secondary_phone')}
-          onChange={accessHandlers('secondary_phone')}
-          value={accessSecondary('secondary_phone')} 
-          placeholder="Phone Number" 
-          isInternational={this.props.isInternational}
-        />
-        {this.props.renderContactTypeField('secondary_type')}
-      </div>
-    );
-  },
-
-});
-
-const CountryPreferences = React.createClass({
-  propTypes: {
-    renderCountryDropdown: React.PropTypes.func,
-  },
-
-  render: function() {
-    return (
-      <div id='country_preferences'>
-        <h3>Country Preferences</h3>
-        <p className="instructions">Please choose 10 United Nations Member States or
-        Observers your school would like to represent. A reference list of
-        countries and their relation to committees is
-        available <a href="http://www.un.org/en/member-states/" target="_blank">online</a>.
-        Please diversify your selection.</p>
-        <ul>
-          {this.props.renderCountryDropdown('01', 'country_pref1')}
-          {this.props.renderCountryDropdown('02', 'country_pref2')}
-          {this.props.renderCountryDropdown('03', 'country_pref3')}
-          {this.props.renderCountryDropdown('04', 'country_pref4')}
-          {this.props.renderCountryDropdown('05', 'country_pref5')}
-          {this.props.renderCountryDropdown('06', 'country_pref6')}
-          {this.props.renderCountryDropdown('07', 'country_pref7')}
-          {this.props.renderCountryDropdown('08', 'country_pref8')}
-          {this.props.renderCountryDropdown('09', 'country_pref9')}
-          {this.props.renderCountryDropdown('10', 'country_pref10')}
-        </ul>
-      </div>
-    );
-  },
-
-});
-
-const SpecialCommitteePreferences = React.createClass({
-  propTypes: {
-    handlers: React.PropTypes.object, 
-    errors: React.PropTypes.object, 
-    specialCommitteePrefValues: React.PropTypes.object, 
-    renderCommittees: React.PropTypes.func, 
-  },
-
-  render: function() {
-    var accessHandlers = _accessSafe.bind(this, this.props.handlers);
-    var accessErrors = _accessSafe.bind(this, this.props.errors);
-    var accessValues = _accessSafe.bind(this, this.props.specialCommitteePrefValues);
-    return (
-      <div id='special_committee_preferences'>
-        <h3>Special Committee Preferences</h3>
-        <p className="instructions">Would your delegation be interested in
-        being represented in the following small/specialized committees?
-        Positions are limited and we may not be able to accommodate all
-        preferences. You can find a reference to our
-        committees <a href="http://www.bmun.org/committees" target="_blank">
-        here</a>.</p>
-        <ul>
-          {this.props.renderCommittees()}
-        </ul>
-        <NumberInput 
-          placeholder="Number of Spanish Speaking Delegates"  
-          onChange={accessHandlers('spanish_speaking_delegates')} 
-          value={accessValues('spanish_speaking_delegates')} 
-        />
-        {accessErrors('spanish_speaking_delegates')}
-        <NumberInput 
-          placeholder="Number of Chinese Speaking Delegates" 
-          onChange={accessHandlers('chinese_speaking_delegates')} 
-          value={accessValues('chinese_speaking_delegates')} 
-        />
-        {accessErrors('chinese_speaking_delegates')}
-      </div>
-    );
-  },
-
-});
-
-const Comments = React.createClass({
-  propTypes: {
-    handler: React.PropTypes.func,
-    value: React.PropTypes.string,
-  },
-
-  render: function() {
-    return (
-      <div id='comments'>
-        <h3>Comments</h3>
-        <p className="instructions">If there are any further details you
-        would like us to know about your participation in BMUN this year or
-        general feedback about the registration process, please comment
-        below.</p>
-        <textarea
-          className="text-input"
-          cols="40"
-          rows="7"
-          onChange={this.props.handler}
-          value={this.props.value}
-        />
-      </div>
-    );
-  },
-
-});
-
-const RegistrationFooter = React.createClass({
-  propTypes: {
-    loading: React.PropTypes.bool,
-  },
-
-  render: function() {
-    return (
-      <div id='registration_footer'>
-        <NavLink direction="left" href="/login">
-          Back to Login
-        </NavLink>
-        <div style={{float: 'right'}}>
-          <span className="help-text"><em>All done?</em></span>
-          <Button
-            color="green"
-            loading={this.props.loading}
-            type="submit">
-            Register
-          </Button>
-        </div>
-      </div>
-    );
-  },
-
 });
 
 module.exports = RegistrationView;
